common/core/geohandlers: rework on mimmo custom binary streams
- mimmo::I/OBinaryStream moved to common module, together with custom
implementation of stream operator for the most important c++ structures
- Binary stream operators designed for mimmo classes as BasicShapes,
MimmoPiercedVector and MimmoNamespace/FileDataInfo moved in the
respective hpp/cpp of their classes
- Simplified InOut files, now carrying only information on PortIn/Out
structures.
- reviewed include chain for BaseManipulation, FVMeshSelection and
MeshSelection (affected by the change)
