+++ src/core/InOut.cpp
@@ -22,172 +22,7 @@
 *
 \*---------------------------------------------------------------------------*/
 #include "InOut.hpp"
-#include "MimmoNamespace.hpp"
 #include "BaseManipulation.hpp"
-
-namespace mimmo{
-    /*!
-        base constructor
-    */
-    IBinaryStream::IBinaryStream(void) : bitpit::IBinaryStream(){};
-    /*!
-        custom constructor
-        @param[in] size of the buffer
-    */
-    IBinaryStream::IBinaryStream(std::size_t size) : bitpit::IBinaryStream(size){};
-    /*!
-        custom constructor
-        @param[in] buffer pointer to buffer
-        @param[in] size size of the buffer
-    */
-    IBinaryStream::IBinaryStream(const char *buffer, std::size_t size) : bitpit::IBinaryStream(buffer,size){};
-    /*!
-        custom constructor
-        @param[in] buffer as vector of char
-    */
-    IBinaryStream::IBinaryStream(const std::vector<char> &buffer): bitpit::IBinaryStream(buffer){};
-
-    /*!
-        base constructor
-    */
-    OBinaryStream::OBinaryStream(): bitpit::OBinaryStream(){};
-    /*!
-        custom constructor
-        @param[in] size size of the buffer
-    */
-    OBinaryStream::OBinaryStream(std::size_t size): bitpit::OBinaryStream(size){};
-};
-
-/*!
-*	Output stream operator for mimmo::ShapeType enum
-*	\param[in] buffer is the output stream
-*	\param[in] var is the element to be streamed
-*	\result Returns the same output stream received in input.
-*/
-mimmo::OBinaryStream& operator<<(mimmo::OBinaryStream  &buffer, const mimmo::ShapeType &var)
-{
-    buffer << static_cast<int> (var);
-    return buffer;
-}
-
-
-/*!
-*	Input stream operator for mimmo::ShapeType enum
-*	\param[in] buffer is the input stream
-*	\param[in] var is the element to be streamed
-*	\result Returns the same input stream received in input.
-*/
-mimmo::IBinaryStream& operator>>(mimmo::IBinaryStream &buffer, mimmo::ShapeType &var)
-{
-    int val;
-    buffer >> val;
-    var = static_cast<mimmo::ShapeType>	(val);
-    return buffer;
-}
-
-/*!
-*	Output stream operator for mimmo::CoordType enum
-*	\param[in] buffer is the output stream
-*	\param[in] var is the element to be streamed
-*	\result Returns the same output stream received in input.
-*/
-mimmo::OBinaryStream& operator<<(mimmo::OBinaryStream  &buffer, const mimmo::CoordType &var)
-{
-    buffer << static_cast<int> (var);
-    return buffer;
-}
-
-
-/*!
-*	Input stream operator for mimmo::CoordType enum
-*	\param[in] buffer is the input stream
-*	\param[in] var is the element to be streamed
-*	\result Returns the same input stream received in input.
-*/
-mimmo::IBinaryStream& operator>>(mimmo::IBinaryStream &buffer, mimmo::CoordType &var)
-{
-    int val;
-    buffer >> val;
-    var = static_cast<mimmo::CoordType>	(val);
-    return buffer;
-}
-
-/*!
-*	Input stream operator for mimmo::FileDataInfo
-*	\param[in] buffer is the input stream
-*	\param[in] var is the element to be streamed
-*	\result Returns the same input stream received in input.
-*/
-mimmo::IBinaryStream& operator>>(mimmo::IBinaryStream &buffer, mimmo::FileDataInfo&  var){
-
-    buffer >> var.ftype;
-    buffer >> var.fdir;
-    buffer >> var.fname;
-
-    return buffer;
-};
-
-/*!
-*	Output stream operator for mimmo::FileDataInfo
-*	\param[in] buffer is the output stream
-*	\param[in] var is the element to be streamed
-*	\result Returns the same output stream received in input.
-*/
-mimmo::OBinaryStream& operator<<(mimmo::OBinaryStream &buffer, const mimmo::FileDataInfo& var){
-
-    buffer<<var.ftype;
-    buffer<<var.fdir;
-    buffer<<var.fname;
-    return buffer;
-
-
-};
-
-/*!
-    Output stream operator for std::string
-    \param[in] buffer is the output stream
-    \param[in] element is the element to be streamed
-    \result Returns the same output stream received in input.
-*/
-mimmo::OBinaryStream& operator<<(mimmo::OBinaryStream &buffer, const std::string& element){
-
-    // Copy until +1 size term, the last one is the end character (old version)
-	// std::vector<char> inputss(element.c_str(), element.c_str()+element.size()+1);
-
-	std::vector<char> inputss(element.c_str(), element.c_str()+element.size());
-    buffer << (std::size_t)inputss.size();
-    for (char & pp: inputss){
-        buffer<<pp;
-    }
-
-    return buffer;
-
-}
-
-/*!
-    Input stream operator for std::string
-    \param[in] buffer is the input stream
-    \param[in] element is the element to be streamed
-    \result Returns the same input stream received in input.
-*/
-mimmo::IBinaryStream& operator>>(mimmo::IBinaryStream &buffer, std::string& element){
-
-	std::size_t nids;
-    buffer >> nids;
-    std::vector<char> inputss(nids);
-    for (std::size_t i = 0; i < nids; ++i){
-        buffer >> inputss[i];
-    }
-
-    // Copy until -1 term, the last one is the end character (old version)
-    //  element = std::string(inputss.begin(), inputss.end()-1);
-
-    element = std::string(inputss.begin(), inputss.end());
-
-
-    return buffer;
-
-}
 
 //==============================================================//
 // DATA TYPE  CLASS	IMPLEMENTATION								//
